Reworking the filter code to look pretty and be faster
Using a helper function to set up the loops, and it is
just SCREAMING. HUGE performance jump in the filters.

Seems to eliminate many of the bounds checks on the
inner loops because the optimizer can tell that all slices
are the same length.

This constraint can't be expressed through the type
system on parameters, but it can totally be expressed in
code and seen by the LLVM optmizer.

Awesome!