Specialize filter_iter using const generic instead of macro
Since real const generics are not implemented yet (RFC2000),
we use the widely used typenum create which fills the gap
quite elegantly.

Monomorphization of this function is now semantically explicit
to Rust and does not anymore rely on an LLVM optimization pass
which might break at anytime in the future.