Expose write_chunk on encoder, bump to 0.3.4
The raw write_chunk function for any arbitrary tag was not
previously exposed at the encoder level in the Rust or C APIs,
which restricted what kind of files you could write.

This should allow both custom chunks and ancillary chunks that
haven't been explicitly supported in the API yet.