Use const generics instead of typenum
This is used in the filter functions to specialize on bits per
pixel, which gives a significant speed boost.

Const generics weren't yet available when mtpng was originally
written, so the typenum crate was used to provide equivalent
semantics with similar enough structure.

Was quite easy to convert this, though it's not a big example
usage. :)